Rename base package for state store committer
