+++ java/statestore-commit/src/main/java/sleeper/statestore/committer/CommitRequestValidationException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.commit;
+package sleeper.statestore.committer;
 
 /**
  * An exception for invalid state store commit requests.
